Create play match fragment
